Moves logic out of app into feature modules.
Other changes:

- moves common logic into a common module
- moves handlers into a handlers module
