+++ src/introductions/mod.rs
@@ -1,2 +1,3 @@
+pub mod commands;
 pub mod queries;
 pub mod voice;
