+++ src/lib.rs
@@ -1,3 +1,6 @@
 pub mod app;
+pub mod common;
 pub mod database_setup;
+pub mod handlers;
 pub mod introductions;
+pub mod rolls;
